cleanup: - rename downloader.js to reflect that it is a service worker - remove useless backup files - remove unused downloader.html - remove unused imports, exports, variables - refactor default configuration into a separate file - simplify random subreddit selection - format some more code with prettier - bugfix "plus.svg" path so it now correctly appears
