--- downloader.html
@@ -1,16 +1,0 @@
-<script>
-
-    alert("oy")
-chrome.runtime.onMessage.addListener(
-  function(arg, sender, sendResponse) {
-    alert("oy")
-    chrome.downloads.download({
-     url: arg,
-    //  filename: saveas,
-    // saveAs: false
-    });
-   }
-);
-
-
-</script>
