+++ manifest.json
@@ -14,7 +14,7 @@
     "newtab": "index.html"
   },
   "background": {
-    "service_worker": "scripts/downloader.js",
+    "service_worker": "scripts/service-worker.js",
     "type": "module"
   },
   "content_security_policy": {
