+++ scripts/default-subreddits.js
@@ -1,0 +1,158 @@
+const curated_subreddits_art = {
+  ArtPorn: {
+    enabled: false,
+  },
+  Art: {
+    enabled: false,
+  },
+  ColorizedHistory: {
+    enabled: false,
+  },
+  SculpturePorn: {
+    enabled: true,
+  },
+  museum: {
+    enabled: false,
+  },
+  Heavymind: {
+    enabled: true,
+  },
+  glitch_art: {
+    enabled: false,
+  },
+  generative: {
+    enabled: false,
+  },
+  pixelsorting: {
+    enabled: false,
+  },
+  nocontextpics: {
+    enabled: false,
+  },
+  creativecoding: {
+    enabled: true,
+  },
+  FractalPorn: {
+    enabled: true,
+  },
+};
+
+const curated_subreddits_imaginary = {
+  ImaginarySkyscapes: {
+    enabled: true,
+  },
+  ImaginaryColorscapes: {
+    enabled: true,
+  },
+  ImaginaryWeather: {
+    enabled: false,
+  },
+  ImaginaryMindscapes: {
+    enabled: true,
+  },
+  ImaginaryLandscapes: {
+    enabled: true,
+  },
+  ImaginaryWildlands: {
+    enabled: false,
+  },
+  ImaginaryCityscapes: {
+    enabled: false,
+  },
+  ImaginaryHellscapes: {
+    enabled: false,
+  },
+  ImaginaryArchers: {
+    enabled: false,
+  },
+  ImaginaryDragons: {
+    enabled: false,
+  },
+  ImaginaryArchers: {
+    enabled: false,
+  },
+  ImaginaryBestOf: {
+    enabled: false,
+  },
+};
+
+const curated_subreddits_nature = {
+  EarthPorn: {
+    enabled: true,
+  },
+  NatureIsFuckingLit: {
+    enabled: true,
+  },
+  BotanicalPorn: {
+    enabled: false,
+  },
+  WaterPorn: {
+    enabled: false,
+  },
+  SeaPorn: {
+    enabled: false,
+  },
+  SkyPorn: {
+    enabled: true,
+  },
+  FirePorn: {
+    enabled: false,
+  },
+  DesertPorn: {
+    enabled: false,
+  },
+  WinterPorn: {
+    enabled: true,
+  },
+  AutumnPorn: {
+    enabled: true,
+  },
+  VillagePorn: {
+    enabled: true,
+  },
+  CityPorn: {
+    enabled: false,
+  },
+  WeatherPorn: {
+    enabled: false,
+  },
+  GeologyPorn: {
+    enabled: false,
+  },
+  SpacePorn: {
+    enabled: true,
+  },
+  BeachPorn: {
+    enabled: false,
+  },
+  MushroomPorn: {
+    enabled: false,
+  },
+  SpringPorn: {
+    enabled: false,
+  },
+  SummerPorn: {
+    enabled: false,
+  },
+  LavaPorn: {
+    enabled: true,
+  },
+};
+
+export const default_subreddits = {
+  creativecoding: {
+    enabled: false,
+  },
+};
+
+export const default_curated_subreddits = {
+  Art: curated_subreddits_art,
+  Imaginary: curated_subreddits_imaginary,
+  Nature: curated_subreddits_nature,
+};
+
+export const default_filters = {
+  sorting: "top",
+  range: "100",
+  time: "all",
+};
